created schemaValidation, extracted classes and all that good stuff
